filter comments in each issue
I added more functionality to the buttons inside each issue for each comment. Now if you create a comment (through an issue) you will be redirected to the comments page displaying all those comments. The same functionality for the other three buttons (r, u, d) - brings you to the comments page that is filtered by the issue you were just on, to do those features to the comments.